+++ src/Backend/Modules/Pages/Engine/Model.php
@@ -153,6 +153,7 @@ class Model
                 $temp['no_follow'] = (bool) (isset($page['meta_data']['seo_follow']) && $page['meta_data']['seo_follow'] == 'nofollow');
                 $temp['hidden'] = (bool) ($page['hidden'] == 'Y');
                 $temp['extra_blocks'] = null;
+                $temp['has_children'] = (bool) ($page['has_children'] == 'Y');
 
                 // any linked extra's?
                 if ($page['extra_ids'] !== null) {
@@ -1228,11 +1229,14 @@ class Model
                  i.id, i.title, i.parent_id, i.navigation_title, i.type, i.hidden, i.data,
                 m.url, m.data AS meta_data,
                 IF(COUNT(e.id) > 0, "Y", "N") AS has_extra,
-                GROUP_CONCAT(b.extra_id) AS extra_ids
+                GROUP_CONCAT(b.extra_id) AS extra_ids,
+                IF(COUNT(p.id), "Y", "N") AS has_children
              FROM pages AS i
              INNER JOIN meta AS m ON i.meta_id = m.id
              LEFT OUTER JOIN pages_blocks AS b ON b.revision_id = i.revision_id AND b.extra_id IS NOT NULL
              LEFT OUTER JOIN modules_extras AS e ON e.id = b.extra_id AND e.type = ?
+             LEFT OUTER JOIN pages AS p ON p.parent_id = i.id AND p.status = "active" AND p.hidden = "N"
+             AND p.language = i.language
              WHERE i.parent_id IN (' . implode(', ', $ids) . ')
                  AND i.status = ? AND i.language = ?
              GROUP BY i.revision_id